pyproject.toml
[project]
name = "local-rag-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "inngest>=0.5.18",
    "ipykernel>=7.2.0",
    "llama-index-core>=0.14.17",
    "llama-index-readers-file>=0.6.0",
    "ollama>=0.6.1",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "python-env>=1.0.0",
    "qdrant-client>=1.17.1",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
]
